cloud-functions/annotation-function/requirements.txt
functions-framework==3.*
opencv-python==4.8.0.76
youtube_dl==2021.12.17
numpy==1.24.4
google-cloud-pubsub==2.18.4
google-cloud-videointelligence==2.11.3
google-cloud-storage==2.10.0
youtube_dl @ git+https://github.com/ytdl-org/youtube-dl.git@master